requirements.txt
asgiref==3.5.0
astroid==2.11.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
decorator==5.1.1
dill==0.3.4
dj-database-url==0.4.2
Django==4.0.2
django-cors-headers==3.11.0
django-google-maps==0.12.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysqlclient==2.1.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pylint==2.13.5
python-decouple==3.6
pytz==2021.3
qrcode==7.3.1
ratelim==0.1.6
razorpay==1.3.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==2.66.0
tomli==2.0.1
twilio==7.6.0
typing-extensions==4.1.1
tzdata==2021.5
Unipath==1.1
urllib3==1.26.8
wrapt==1.14.0
